Remove Proc Macro hack Since Rust 1.4.5 procedural macros have now been stablized in the expression position. This allows us to remove the proc_macro_hack that was being used to emulate this feature.
