--- macro_impl/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "rust_decimal_macro_impls"
-version = "1.7.0"
-authors = ["Paul Mason <paul@form1.co.nz>"]
-edition = "2018"
-description = "Shorthand macros to assist creating Decimal types. Do not depend on this directly; use rust_decimal_macros"
-documentation = "https://docs.rs/rust_decimal/"
-repository = "https://github.com/paupino/rust-decimal"
-readme = "../README.md"
-keywords = ["decimal","financial","fixed","precision"]
-categories = ["science","data-structures"]
-license = "MIT"
-
-[dependencies]
-rust_decimal = { path = "../", version = "1.7.0" }
-proc-macro-hack = "0.5"
-quote = "1.0"
-
-[lib]
-proc-macro = true
