+++ macros/Cargo.toml
@@ -13,5 +13,7 @@ license = "MIT"
 
 [dependencies]
 rust_decimal = { path = "..", version = "1.7.0" }
-rust_decimal_macro_impls = { path = "../macro_impl", version = "1.7.0" }
-proc-macro-hack = "0.5"
+quote = "1.0"
+
+[lib]
+proc-macro = true
